+++ packages/map/src/index.ts
@@ -1,5 +1,6 @@
 import { run, subcommands } from 'cmd-ts';
 
+import { deployCommand } from './cli/action.deploy.ts';
 import { downloadCommand } from './cli/action.download.ts';
 import { listMapSheetsCommand } from './cli/action.list.mapsheet.ts';
 import { ProduceCommand } from './cli/action.produce.ts';
@@ -11,6 +12,7 @@ const Cli = subcommands({
     produce: ProduceCommand,
     'list-mapsheets': listMapSheetsCommand,
     download: downloadCommand,
+    deploy: deployCommand,
   },
 });