+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## [0.4.0](https://github.com/linz/topographic-system/compare/v0.3.0...v0.4.0) (2026-01-13)
+
+
+### Features
+
+* **map:** Generate png files from the qgis project for screenshots. BM-1349 ([#20](https://github.com/linz/topographic-system/issues/20)) ([2e9b458](https://github.com/linz/topographic-system/commit/2e9b458a6b767fe8e409c0e82f97b9a452b6d595))
+* Visual diff in PR BM-1424 ([#21](https://github.com/linz/topographic-system/issues/21)) ([9f0e004](https://github.com/linz/topographic-system/commit/9f0e00458954daa2f7bc06cf35420cd57f14a7a5))
+
+
+### Bug Fixes
+
+* init arch keyring BM-1474 ([#26](https://github.com/linz/topographic-system/issues/26)) ([b42dfef](https://github.com/linz/topographic-system/commit/b42dfefd378691fc1c04857e43360ea0dca670f5))
+
+## [0.3.0](https://github.com/linz/topographic-system/compare/v0.2.0...v0.3.0) (2026-01-05)
+
+
+### Features
+
+* **kart:** extract data from kart repo to gpkg BM-1447 ([#17](https://github.com/linz/topographic-system/issues/17)) ([0e13c2f](https://github.com/linz/topographic-system/commit/0e13c2f37f460ceee0eb2a6d6403e1e13abdfef4))
+* **map:** Add Download cli and list mapsheet clis for argo workflow. BM-1393 ([#11](https://github.com/linz/topographic-system/issues/11)) ([1eb12a6](https://github.com/linz/topographic-system/commit/1eb12a6603ed838ff321d6763985475a36902bcc))
+* **map:** cli for deploying qgis project into aws. BM-1394 ([#15](https://github.com/linz/topographic-system/issues/15)) ([9b5378e](https://github.com/linz/topographic-system/commit/9b5378ed8659dfbb86ffc8be64a05a13c2910abf))
+* **map:** Create Stac files for the output files and simple tiff validation. BM-1392 ([#9](https://github.com/linz/topographic-system/issues/9)) ([88ebf31](https://github.com/linz/topographic-system/commit/88ebf313661b7ab0389daebbed3fad8b5f119392))
+* python configuration and CI BM-1456 ([#12](https://github.com/linz/topographic-system/issues/12)) ([33fab64](https://github.com/linz/topographic-system/commit/33fab642ac5e2565847a033c6e174f85b34c8167))
+
+
+### Bug Fixes
+
+* set packages and id-token write permission BM-1464 ([#19](https://github.com/linz/topographic-system/issues/19)) ([9add832](https://github.com/linz/topographic-system/commit/9add83248234a02fe6657e4bb5e43febb4204a2b))
+
 ## [0.2.0](https://github.com/linz/topographic-system/compare/v0.1.0...v0.2.0) (2025-12-01)